Refactor API calls into charity service
